# Installer cleanup